Chunk 4: extract findings/NearNameMatcher from SuggestionsRenderer
SuggestionsRenderer mixed near-name scoring with Markdown formatting. Move all
scoring — similarity/levenshtein/token signals, the candidate-pool prep, the
git-rename/near-name Hints, the config-key match scan, uniqueBasenameDoc, and
packageMentionLines — into a new findings/NearNameMatcher. The renderer keeps
only composition (render/composeBullets/configKeyBullets/strandedProse) and
formats the matcher's ranked output. configKeyBullets now reads m.keyName()
for the exact-match bullet (equal to the gone property there), dropping the
renderer's private lastDotSegment. The direct scoring unit test moves to
findings/NearNameMatcherTest. No behavior change; run dir byte-identical to
the golden.
